+++ tests/testthat/test-chat.R
@@ -1,10 +1,9 @@
 test_that("can get and set the system prompt", {
-  chat <- chat_openai_test(
-    turns = list(
-      Turn("user", "Hi"),
-      Turn("assistant", "Hello")
-    )
-  )
+  chat <- chat_openai_test()
+  chat$set_turns(list(
+    Turn("user", "Hi"),
+    Turn("assistant", "Hello")
+  ))
 
   # NULL -> NULL
   chat$set_system_prompt(NULL)
@@ -189,24 +188,23 @@ test_that("can retrieve tokens with or without system prompt", {
 
 test_that("has a basic print method", {
   chat <- chat_openai(
-    "You're a helpful assistant that returns very minimal output",
-    turns = list(
-      Turn("user", "What's 1 + 1?\nWhat's 1 + 2?"),
-      Turn("assistant", "2\n\n3", tokens = c(15, 5))
-    )
+    "You're a helpful assistant that returns very minimal output"
   )
+  chat$set_turns(list(
+    Turn("user", "What's 1 + 1?\nWhat's 1 + 2?"),
+    Turn("assistant", "2\n\n3", tokens = c(15, 5))
+  ))
   expect_snapshot(chat)
 })
 
 test_that("print method shows cumulative tokens & cost", {
-  chat <- chat_openai(
-    turns = list(
-      Turn("user", "Input 1"),
-      Turn("assistant", "Output 1", tokens = c(15000, 500)),
-      Turn("user", "Input 2"),
-      Turn("assistant", "Output 1", tokens = c(30000, 1000))
-    )
-  )
+  chat <- chat_openai()
+  chat$set_turns(list(
+    Turn("user", "Input 1"),
+    Turn("assistant", "Output 1", tokens = c(15000, 500)),
+    Turn("user", "Input 2"),
+    Turn("assistant", "Output 1", tokens = c(30000, 1000))
+  ))
   expect_snapshot(chat)
 
   expect_equal(chat$get_cost(), dollars(0.1275))